Move additional specification to a separate module
